# MissingTypehintCheck: reduce duplicate work